backend/requirements.txt
# ============================================
# Runtime & Flask Extensions
# ============================================
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-SocketIO==5.3.6

# ============================================
# Persistence, Messaging & Validation
# ============================================
psycopg2-binary==2.9.9      # PostgreSQL adapter (optional in dev)
redis==5.0.1                # Rate limiting, cache, Celery broker
marshmallow==3.20.1         # Schema validation & DTOs

# ============================================
# Async Workloads & Scheduling
# ============================================
celery==5.3.4
APScheduler==3.10.4
python-socketio==5.10.0
eventlet==0.34.2            # SocketIO async worker

# ============================================
# API Documentation & DX
# ============================================
flasgger==0.9.7.1
flask-swagger-ui==4.11.1
python-dotenv==1.0.0

# ============================================
# Observability & Logging
# ============================================
structlog==24.1.0
python-json-logger==2.0.7

# ============================================
# Deployment
# ============================================
gunicorn==21.2.0

# ============================================
# Testing & Quality Gates
# ============================================
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
